requirements.txt
langchain-google-genai
google-generativeai
requests